correct deployment scripts

+++ scripts/stop_server.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
-pkill $(cat /deployment/pid.file)
-
-echo "Killed process at: "$(cat /deployment/pid.file)
+echo "Trying to Stop the Server (if its running)"
+if [ -d "/deployment/pid.file" ]; then
+    pkill $(cat /deployment/pid.file);
+    echo "Killed process at: "$(cat /deployment/pid.file)
+fi
